# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeNexus - LeetCode-style coding platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProblemSummaryDifficulty = typeof ProblemSummaryDifficulty[keyof typeof ProblemSummaryDifficulty];


export const ProblemSummaryDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export interface ProblemSummary {
  id: number;
  title: string;
  difficulty: ProblemSummaryDifficulty;
  tags: string[];
  acceptanceRate: number;
  totalSubmissions: number;
  isSolved: boolean;
  isPremium?: boolean;
}

export type ProblemDifficulty = typeof ProblemDifficulty[keyof typeof ProblemDifficulty];


export const ProblemDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export type ProblemStarterCode = {[key: string]: string};

export interface ProblemExample {
  input: string;
  output: string;
  /** @nullable */
  explanation?: string | null;
}

export interface Problem {
  id: number;
  title: string;
  difficulty: ProblemDifficulty;
  tags: string[];
  description: string;
  examples: ProblemExample[];
  constraints: string[];
  hints?: string[];
  starterCode: ProblemStarterCode;
  acceptanceRate: number;
  totalSubmissions: number;
  isSolved: boolean;
  companies?: string[];
  /** @nullable */
  editorial?: string | null;
}

export type ProblemInputDifficulty = typeof ProblemInputDifficulty[keyof typeof ProblemInputDifficulty];


export const ProblemInputDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export interface ProblemInput {
  title: string;
  difficulty: ProblemInputDifficulty;
  description: string;
  tags: string[];
  constraints?: string[];
  hints?: string[];
  companies?: string[];
}

export interface ProblemStats {
  total: number;
  easy: number;
  medium: number;
  hard: number;
  solvedEasy: number;
  solvedMedium: number;
  solvedHard: number;
}

export type SubmissionVerdict = typeof SubmissionVerdict[keyof typeof SubmissionVerdict];


export const SubmissionVerdict = {
  Accepted: 'Accepted',
  Wrong_Answer: 'Wrong Answer',
  Time_Limit_Exceeded: 'Time Limit Exceeded',
  Memory_Limit_Exceeded: 'Memory Limit Exceeded',
  Compilation_Error: 'Compilation Error',
  Runtime_Error: 'Runtime Error',
} as const;

export interface Submission {
  id: number;
  problemId: number;
  userId: number;
  language: string;
  code: string;
  verdict: SubmissionVerdict;
  /** @nullable */
  runtime: number | null;
  /** @nullable */
  memory: number | null;
  /** @nullable */
  testCasesPassed?: number | null;
  /** @nullable */
  totalTestCases?: number | null;
  createdAt: string;
  /** @nullable */
  problemTitle?: string | null;
  /** @nullable */
  username?: string | null;
}

export interface SubmissionInput {
  problemId: number;
  language: string;
  code: string;
  userId?: number;
}

export interface RunCodeInput {
  problemId: number;
  language: string;
  code: string;
  customInput?: string;
}

export interface RunResult {
  output: string;
  status: string;
  /** @nullable */
  runtime?: number | null;
  /** @nullable */
  error?: string | null;
}

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  rating: number;
  rank: number;
  solvedCount: number;
  streak?: number;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
  badges?: string[];
}

export interface UserInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  bio?: string;
}

export interface LanguageStat {
  language: string;
  count: number;
}

export interface UserStats {
  userId: number;
  solvedEasy: number;
  solvedMedium: number;
  solvedHard: number;
  totalSubmissions: number;
  acceptedSubmissions: number;
  streak: number;
  rating: number;
  languageStats?: LanguageStat[];
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthResponse {
  user: UserProfile;
  token: string;
}

export type ContestStatus = typeof ContestStatus[keyof typeof ContestStatus];


export const ContestStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  finished: 'finished',
} as const;

export interface Contest {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  startTime: string;
  endTime: string;
  status: ContestStatus;
  participantCount: number;
  problemCount: number;
}

export type ContestDetailStatus = typeof ContestDetailStatus[keyof typeof ContestDetailStatus];


export const ContestDetailStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  finished: 'finished',
} as const;

export interface ContestDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  startTime: string;
  endTime: string;
  status: ContestDetailStatus;
  participantCount: number;
  problems: ProblemSummary[];
}

export interface ContestInput {
  title: string;
  description?: string;
  startTime: string;
  endTime: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  score: number;
  solvedCount: number;
  penaltyTime: number;
}

export interface DashboardSummary {
  totalProblems: number;
  totalUsers: number;
  totalSubmissions: number;
  totalAccepted: number;
  activeContests: number;
  easyCount?: number;
  mediumCount?: number;
  hardCount?: number;
}

export interface ActivityItem {
  id: number;
  userId: number;
  username: string;
  problemId: number;
  problemTitle: string;
  verdict: string;
  language: string;
  createdAt: string;
}

export type ListProblemsParams = {
difficulty?: ListProblemsDifficulty;
tag?: string;
search?: string;
status?: ListProblemsStatus;
};

export type ListProblemsDifficulty = typeof ListProblemsDifficulty[keyof typeof ListProblemsDifficulty];


export const ListProblemsDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export type ListProblemsStatus = typeof ListProblemsStatus[keyof typeof ListProblemsStatus];


export const ListProblemsStatus = {
  solved: 'solved',
  attempted: 'attempted',
  unsolved: 'unsolved',
} as const;

export type ListSubmissionsParams = {
problemId?: number;
userId?: number;
limit?: number;
};

export type ListUsersParams = {
limit?: number;
};

export type ListContestsParams = {
status?: ListContestsStatus;
};

export type ListContestsStatus = typeof ListContestsStatus[keyof typeof ListContestsStatus];


export const ListContestsStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  finished: 'finished',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};

export type GetGlobalLeaderboardParams = {
limit?: number;
};