# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeNexus - LeetCode-style coding platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /problems:
    get:
      operationId: listProblems
      tags: [problems]
      summary: List all problems with optional filters
      parameters:
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [Easy, Medium, Hard]
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [solved, attempted, unsolved]
      responses:
        "200":
          description: List of problems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemSummary"
    post:
      operationId: createProblem
      tags: [problems]
      summary: Create a new problem (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProblemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /problems/{id}:
    get:
      operationId: getProblem
      tags: [problems]
      summary: Get a single problem with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Problem detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not found
...
  /contests:
    get:
      operationId: listContests
      tags: [contests]
      summary: List all contests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, ongoing, finished]
      responses:
        "200":
          description: List of contests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contest"
    post:
      operationId: createContest
      tags: [contests]
      summary: Create a new contest (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContestInput"
      responses:
        "201":
          description: Contest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"
  /contests/{id}:
    get:
      operationId: getContest
      tags: [contests]
      summary: Get a single contest with problems
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contest detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContestDetail"
        "404":
          description: Not found
  /contests/{id}/leaderboard:
    get:
      operationId: getContestLeaderboard
      tags: [contests]
      summary: Get leaderboard for a contest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contest leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  # ── Dashboard ─────────────────────────────────────────────
...
components:
  schemas:
...
    ProblemSummary:
      type: object
      required: [id, title, difficulty, tags, acceptanceRate, totalSubmissions, isSolved]
      properties:
        id:
          type: integer
        title:
          type: string
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        tags:
          type: array
          items:
            type: string
        acceptanceRate:
          type: number
        totalSubmissions:
          type: integer
        isSolved:
          type: boolean
        isPremium:
          type: boolean
    Problem:
      type: object
      required: [id, title, difficulty, tags, description, examples, constraints, starterCode, acceptanceRate, totalSubmissions, isSolved]
      properties:
        id:
          type: integer
        title:
          type: string
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        tags:
          type: array
          items:
            type: string
        description:
          type: string
        examples:
          type: array
          items:
            $ref: "#/components/schemas/ProblemExample"
        constraints:
          type: array
          items:
            type: string
        hints:
          type: array
          items:
            type: string
        starterCode:
          type: object
          additionalProperties:
            type: string
        acceptanceRate:
          type: number
        totalSubmissions:
          type: integer
        isSolved:
          type: boolean
        companies:
          type: array
          items:
            type: string
        editorial:
          type: ["string", "null"]
    ProblemExample:
      type: object
      required: [input, output]
      properties:
        input:
          type: string
        output:
          type: string
        explanation:
          type: ["string", "null"]
    ProblemInput:
      type: object
      required: [title, difficulty, description, tags]
      properties:
        title:
          type: string
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        hints:
          type: array
          items:
            type: string
        companies:
          type: array
          items:
            type: string
...
    Contest:
      type: object
      required: [id, title, startTime, endTime, status, participantCount, problemCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, finished]
        participantCount:
          type: integer
        problemCount:
          type: integer
    ContestDetail:
      type: object
      required: [id, title, startTime, endTime, status, participantCount, problems]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, finished]
        participantCount:
          type: integer
        problems:
          type: array
          items:
            $ref: "#/components/schemas/ProblemSummary"
    ContestInput:
      type: object
      required: [title, startTime, endTime]
      properties:
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
        endTime:
          type: string
    LeaderboardEntry:
      type: object
      required: [rank, userId, username, score, solvedCount, penaltyTime]
      properties:
        rank:
          type: integer
        userId:
          type: integer
        username:
          type: string
        score:
          type: integer
        solvedCount:
          type: integer
        penaltyTime:
          type: integer

    # ── Dashboard schemas ─────────────────────────────────────
...