# lib/api-zod/src/generated/types/contestDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeNexus - LeetCode-style coding platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ContestDetailStatus } from './contestDetailStatus';
import type { ProblemSummary } from './problemSummary';

export interface ContestDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  startTime: string;
  endTime: string;
  status: ContestDetailStatus;
  participantCount: number;
  problems: ProblemSummary[];
}