# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeNexus - LeetCode-style coding platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemDifficulty } from './problemDifficulty';
import type { ProblemExample } from './problemExample';
import type { ProblemStarterCode } from './problemStarterCode';

export interface Problem {
  id: number;
  title: string;
  difficulty: ProblemDifficulty;
  tags: string[];
  description: string;
  examples: ProblemExample[];
  constraints: string[];
  hints?: string[];
  starterCode: ProblemStarterCode;
  acceptanceRate: number;
  totalSubmissions: number;
  isSolved: boolean;
  companies?: string[];
  /** @nullable */
  editorial?: string | null;
}