# lib/api-zod/src/generated/types/problemSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeNexus - LeetCode-style coding platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemSummaryDifficulty } from './problemSummaryDifficulty';

export interface ProblemSummary {
  id: number;
  title: string;
  difficulty: ProblemSummaryDifficulty;
  tags: string[];
  acceptanceRate: number;
  totalSubmissions: number;
  isSolved: boolean;
  isPremium?: boolean;
}